image-resizer/requirements.txt
flask
minio
pillow
